backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
google-genai==1.0.0
python-dotenv==1.0.1
aiofiles==24.1.0
pydantic==2.10.0
pytest==8.3.0
pytest-asyncio==0.24.0
httpx==0.27.0
